LayoutTests/http/tests/resourceLoadStatistics/strip-referrer-to-origin-for-prevalent-subresource-redirects.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
    description("Tests that only the origin is sent as referrer in redirects to prevalent resources without user interaction.");
    jsTestIsAsync = true;
    testRunner.dumpChildFramesAsText();

    function setEnableFeature(enable, completionHandler) {
        internals.setResourceLoadStatisticsEnabled(enable);
        testRunner.setCookieStoragePartitioningEnabled(enable);
        if (!enable)
            testRunner.statisticsResetToConsistentState(completionHandler);
        else
            completionHandler();
    }

    function openIframe(url, onLoadHandler) {
        const element = document.createElement("iframe");
        element.src = url;
        if (onLoadHandler) {
            element.onload = onLoadHandler;
        }
        document.body.appendChild(element);
    }

    var referrer;
    setEnableFeature(true, function() {
        if (testRunner.isStatisticsPrevalentResource("http://localhost"))
            testFailed("Localhost was classified as prevalent resource before the test started.");

        fetch("resources/echo-referrer.php").then(function(response) {
            return response.text();
        }).then(function(data) {
            referrer = data;
            shouldBeEqualToString("referrer", "http://127.0.0.1:8000/resourceLoadStatistics/strip-referrer-to-origin-for-prevalent-subresource-redirects.html");

            testRunner.setStatisticsPrevalentResource("http://localhost", true, function() {
                if (!testRunner.isStatisticsPrevalentResource("http://localhost"))
                    testFailed("Host did not get set as prevalent resource.");

                testRunner.statisticsUpdateCookiePartitioning(function() {
                    openIframe("resources/redirect.php?redirectTo=http://localhost:8000/resourceLoadStatistics/resources/echo-referrer.php", function() {
                        setEnableFeature(false, finishJSTest);
                    });
                });
            });
        }).catch(function(error) {
            console.log(error.message);
            setEnableFeature(false, finishJSTest);
        });
    });
</script>
</body>
</html>
